chore: fix compiler warnings and update ARCHITECTURE.md
- Add copy assignment operators to fix -Wdeprecated-copy warnings
  (NextAction, PositionInfo, CraftData, UnitPosition)
- Fix member initialization order to match declaration order
- Fix sign comparison warnings in loop counters
- Remove pessimizing std::move() calls
- Add [[maybe_unused]] to intentionally unused parameters
- Update ARCHITECTURE.md with current statistics:
  - 333 files changed, +18,112/-2,516 lines, 28 commits
  - BotItemService expanded to ~780 lines
  - Module now compiles warning-free

+++ src/Ai/Base/Trigger/GenericTriggers.cpp
@@ -169,7 +169,7 @@ bool BuffTrigger::IsActive()
     Aura* aura = botAI->GetServices().GetSpellService().GetAura(spell, target, checkIsOwner, checkDuration);
     if (!aura)
         return true;
-    if (beforeDuration && aura->GetDuration() < beforeDuration)
+    if (beforeDuration && aura->GetDuration() < static_cast<int32>(beforeDuration))
         return true;
     return false;
 }
@@ -415,7 +415,7 @@ bool HealerShouldAttackTrigger::IsActive()
     return true;
 }
 
-bool ItemCountTrigger::IsActive() { return AI_VALUE2(uint32, "item count", item) < count; }
+bool ItemCountTrigger::IsActive() { return static_cast<int32>(AI_VALUE2(uint32, "item count", item)) < count; }
 
 bool InterruptSpellTrigger::IsActive()
 {
